Fix icons by autoprefixing the mask-* CSS properties
mask-image and other mask-* properties only became supported
unprefixed in very recent Chrome versions. Before that only
-webkit-mask-* worked.
The (embedded) Chromium in OBS seems to be at least somewhat older.

To fix this, we now use the PostCSS 'autoprefixer' plugin.
We make sure to include the last 2 years of Chrome versions
in the newly added .bowserslistrc, so older Chrome in OBS is covered.
We also include the "defaults" browsers so other relevant browsers
are also supported.