Split out the user-defined theme and fixed CSS
Users will usually only want to edit the CSS variables.
Thus we split out that part (in public/theme.css) and
only optimize the rest (now in /style.css) through Vite/rollup.

We duplicate the default CSS variable values in styles.css,
so users can selectively override values they want to change.

We use CSS @layer to put the animate CSS and the poll tool default CSS
into layers, which will always lose to styles from the default layer,
which we use for the theme.css. This way it is easy for users
to override styles (especially the CSS variables).
We also don't have to pay attention to CSS inclusion order for
style.css/theme.css, as this does not matter across layers.

We are also moving the import of the animate.css library to style.css
as this is a central place. This way it is easier to see that it lives
in the 'libs' layer.