Introduce JS config & Inline SVG icons
Use an (optional) window.config to enable (default) or disable icons for
win/tie-break. We also remove the respective CSS variables, as the icons
are now fixed.

Inline the SVG icons using vite-svg-loader, so they don't have to be
loaded separately. This also simplifies the CSS a bit, as we don't have
to rely on mask-* properties, but can directly use 'color' to change the
icon color.

Also:
Moving the icon files from /public and /resources, so they can be
imported in JS.
Reformatting some files with Prettier.