Remove dependabot.yml
Currently there is no big need to get automatic package updates. Also
the current (default) configuration of dependabot was creating PRs for
major updates, which is not what we want for the time being as it quite
quickly leads to problems with node version 16 that we are bound to by
glitch.com.